requirements.txt
aiosmtplib==1.1.7
anyio==3.6.2
bcrypt==4.0.1
beanie==1.11.0
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
fastapi-jwt-auth==0.5.0
fastapi-mail==1.2.2
h11==0.14.0
httptools==0.5.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==2.5.1
multidict==6.0.3
passlib==1.7.4
pydantic==1.10.2
PyJWT==1.7.1
pymongo==3.13.0
pyotp==2.8.0
python-decouple==3.3
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.2
